test: add support for Houdini 20.0
To do this (and make future versions easier) the testing is converted to a more Houdin version-centric
approach rather than the Python version. This has some additional requirements:

Local Testing:

For local testing, the tox targets have been renamed from using the standard Python names (py39, py310, etc)
to the Houdini versions (h195, h20). Since I use Houdini via rez, in order to select a version I need to
pass along a version qualifier to ensure that the correct version is launched so this is now accomplished
via an extra env var (defaulting to empty) that the Houdini specific targets can set and will be added
to the hython command. Since the Github targets rely on Houdini already being available and set up they
will not care about this and will function as expected.

Github Actions:

For Github, the version is now chosen via a matrix on the major.minor version. As specific versions of
Python are still needed, we specify that via matching include fields. We also need to add a new 3.10
gh-actions tox entry to execute with py310.

We also bump the version of install-houdini-apprentice-license-action to 5 in order to no longer need
to specify a Houdini version for the license. This change necessitates that we ensure the setup-python
action is called before the license action as that is no longer setting up the python that it needs.

fixes: #11